+++ tests/test_revoke.py
@@ -1,9 +1,11 @@
+from unittest.mock import patch
+
 import pytest
 from sqlalchemy import Column, Integer, MetaData, Sequence, Table, table
 from sqlalchemy.dialects import postgresql
 from sqlalchemy.ext.declarative import declarative_base
 
-from pg_grant import PgObjectType
+from pg_grant import PgObjectType, Privileges
 from pg_grant.sql import revoke
 
 meta = MetaData()
@@ -147,3 +149,29 @@ def test_compile_revoke_invalid(type, target, schema, arg_types):
 def test_compile_revoke_privs_invalid(privs):
     with pytest.raises(ValueError):
         str(revoke(privs, PgObjectType.TABLE, 't', 'alice'))
+
+
+@pytest.mark.parametrize('grantee, privs, privswgo, type_, target, kw', [
+    ('alice', ['SELECT', 'INSERT'], [], PgObjectType.FOREIGN_TABLE, 'table1', {}),
+    ('alice', ['SELECT'], ['INSERT'], PgObjectType.TABLE, 'table2', {'schema': 's'}),
+])
+def test_privileges_as_revoke_statements(grantee, privs, privswgo, type_, target, kw):
+    priv = Privileges(grantee, 'bob', privs, privswgo)
+
+    def fake_revoke(*args, **kwargs):
+        return args, kwargs
+
+    patch_revoke = patch('pg_grant.sql.revoke', fake_revoke)
+
+    with patch_revoke:
+        r = priv.as_revoke_statements(type_, target, **kw)
+
+    expected = []
+
+    if privs:
+        expected.append(((privs, type_, target, grantee), kw))
+
+    if privswgo:
+        expected.append(((privswgo, type_, target, grantee), kw))
+
+    assert r == expected